+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - 2025-12-10
+
+### Added
+
+- Automatic config migration on plugin startup to add new config keys
+- Config versioning system with backup creation before migration
+- Whitelist management can be disabled via `whitelist.enabled` config option
+
 ## [0.15.0] - 2025-12-10
 
 ### Added